requirements.txt
streamlit
python-docx
python-pptx
PyPDF2
langchain
langchain_community
langchain_google_genai
langchain_text_splitters
sentence-transformers
faiss-cpu
cohere
